refactor(app): lift the send pipeline out of ChatPanel
`ChatPanel` was 2,316 lines, and roughly a third of it was the path between
"the user pressed enter" and "the message is on its way to a session" —
`sendIntoSession`, `handleSubmit`, and `createSessionAndSendFirst`, three
mutually-recursive functions with nothing to do with laying out a chat pane.

They move verbatim into `useChatSend`. The coupling turned out to be far
lighter than the size suggests: 705 lines reaching back into the component
for 15 values, all of them component-local state, refs, or derived values
that genuinely belong to the caller. `parseSlashToken` goes with them —
nothing else used it.

The functions stay plain (not `useCallback`), exactly as they were, so
their re-creation-per-render semantics are unchanged.

ChatPanel 2,316 -> 1,578 lines. 33 now-unused imports pruned from it, and
the hook carries only the 58 it actually needs.
